==> Constructing_ML_Model/4_Creating_Final_Dataset.ipynb <==
# Creating Final Dataset and save in to Comma Seperated Value(.csv) Files

## Creating Final Data Set
In here, we concatenate phishing1.csv, phishing2.csv, legitimate1.csv and legitimate1.csv and make a final datatset.csv file

Creating DataFrame for Uploaded Phishing URLs Files

#importing required packages for this module
import pandas as pd

phishing = pd.read_csv("CSV FILES/5-phish_features.csv")
phishing.head()

Creating DataFrame for Uploaded Phishing URLs Files

legitimate = pd.read_csv("CSV FILES/6-legitimate_features.csv")
legitimate.head()

## **5. Final Dataset**
In the above section we formed two dataframes of legitimate & phishing URL features. Now, we will combine them to a single dataframe and export the data to csv file for the Machine Learning training done in other notebook.

#Concatenating the dataframes into one 
urldata = pd.concat([legitimate, phishing]).reset_index(drop=True)
urldata.head()

# Storing the data in CSV file
urldata.to_csv('7-Final_dataset.csv', index=False)

### Contributors - Deshitha Hansajith [Linked In](https://www.linkedin.com/in/deshitha-hansajith/), Nipuni Dilukshika [Linked In](https://www.linkedin.com/in/nipuni-dilukshika-186764197/)
